+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double height = 1.87; // рост в метрах
         double weight = 98; //вес в килограммах
-        int bmi = (int) service.calculate(height, weight);
+        int bmi = service.calculate(height, weight);
         System.out.println(bmi);
     }
 }
